Move 2 .h files to include
