+++ trezor-crypto/include/TrezorCrypto/zkp_schnorr.h
